refactor: group routing module files into dirs by feature
