+++ projects/ngx-meta/src/routing/index.ts
@@ -1,5 +1,2 @@
-// Main providers
-export * from './src/ngx-meta-routing.module'
-export * from './src/provide-ngx-meta-routing'
-// Others
-export * from './src/ngx-meta-route-data'
+export * from './src/providers'
+export * from './src/route-metadata'
