+++ projects/ngx-meta/src/routing/src/listener/router-listener.spec.ts
@@ -3,7 +3,7 @@ import { TestBed } from '@angular/core/testing'
 import { MockProvider } from 'ng-mocks'
 import { EventType, Router, RouterEvent } from '@angular/router'
 import { EventEmitter } from '@angular/core'
-import { enableAutoSpy } from '@davidlj95/ngx-meta/__tests__/enable-auto-spy'
+import { enableAutoSpy } from '@/ngx-meta/test/enable-auto-spy'
 import { NgxMetaService } from '@davidlj95/ngx-meta/core'
 import { ROUTER_LISTENER, RouterListener } from './router-listener'
 import { Observable } from 'rxjs'
