+++ projects/ngx-meta/src/routing/src/providers/index.ts
@@ -1,0 +1,2 @@
+export { NgxMetaRoutingModule } from './ngx-meta-routing.module'
+export { provideNgxMetaRouting } from './provide-ngx-meta-routing'
