+++ projects/ngx-meta/src/routing/src/route-metadata/index.ts
@@ -1,0 +1,1 @@
+export { NgxMetaRouteData } from './ngx-meta-route-data'
