+++ projects/ngx-meta/src/routing/src/route-metadata/ngx-meta-route-data.ts
